Refactored a bit, broke up read/write pulse funcitonality into multiple modules
