+++ deploy_test.txt
@@ -2,8 +2,7 @@
 # To upload files onto a pico on port `/dev/ttyACM0` run:
 # rshell -p /dev/ttyACM0 --buffer-size 512 -f upload.txt
 
-mkdir /pyboard/tesla_cooler  # will fail if dir already exists, this is fine.
-cp ./tesla_cooler/* /pyboard/tesla_cooler
+rsync -m ./tesla_cooler/ /pyboard/tesla_cooler
 cp ./*.py /pyboard/
 
-repl ~ from test import test ~ test()
+repl ~ from main_test import main ~ main()
