+++ main_prod.py
@@ -147,7 +147,3 @@ def main() -> None:
     )
 
     print("Timers configured!")
-
-
-if __name__ == "__main__":
-    main()
