+++ tools/apply_black.sh
@@ -7,4 +7,4 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd ${DIR}/..
 source ./venv/bin/activate
 
-LC_ALL=C.UTF-8 LANG=C.UTF-8 black ./main.py . -l 100 --exclude=venv --exclude=.micropy
+LC_ALL=C.UTF-8 LANG=C.UTF-8 black ./*.py . -l 100 --exclude=venv --exclude=.micropy
