+++ tools/run_black.sh
@@ -8,4 +8,4 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd ${DIR}/..
 source ./venv/bin/activate
 
-LC_ALL=C.UTF-8 LANG=C.UTF-8 black . main.py tesla_cooler/ test/ -l 100 --diff --check --exclude=venv
+LC_ALL=C.UTF-8 LANG=C.UTF-8 black . *.py tesla_cooler/ test/ -l 100 --diff --check --exclude=venv
